requirements.txt
streamlit
pandas
sqlite3  
scikit-learn
numpy
joblib
openai==0.28.1
python-dotenv
os